Сделала 4 задание
